requirements.txt
rl-games==1.1.4
gym==0.26.2
numpy==1.21.6
termcolor
matplotlib
trimesh==4.4.1
rtree==1.3.0
torchgeometry==0.1.2
joblib
tqdm
smplx==0.1.28
open3d==0.18.0
wandb>=0.15.0
# dependencies for scenepic (see https://github.com/microsoft/scenepic/blob/main/requirements.txt)
scenepic==1.1.0
opencv-python==4.10.0.84
nbstripout
pytest-emoji
pytest-md
pytest-cov
pytest
requests
twine
wheel